% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floaties-package.R
\docType{package}
\name{floaties-package}
\alias{floaties}
\alias{floaties-package}
\title{floaties: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
