% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{floaties_palette}
\alias{floaties_palette}
\title{Colours shared by every engine}
\usage{
floaties_palette(which = names(FLOATIES_COLOURS))
}
\arguments{
\item{which}{Name of a colour: \code{"bar"} or \code{"rule"}.}
}
\value{
A hex colour string.
}
\description{
Colours shared by every engine
}
\examples{
floaties_palette("bar")

}
