% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{new_engine}
\alias{new_engine}
\title{Build an engine object}
\usage{
new_engine(subclass, ...)
}
\arguments{
\item{subclass}{A string naming the engine, e.g. \code{"ggplot2"}.}

\item{...}{Engine-specific options.}
}
\value{
An object of class \code{floaties_engine}.
}
\description{
The constructor an external engine package calls. Given \code{subclass = "foo"},
the returned object has class \code{c("floaties_foo", "floaties_engine")}, so
\code{render_waterfall.floaties_foo()} will be dispatched to.
}
\examples{
new_engine("text", width = 40)

}
