% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimlane.R
\name{validate_swimlane_data}
\alias{validate_swimlane_data}
\title{Validate canonical swimlane data}
\usage{
validate_swimlane_data(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
\code{x}, invisibly. Errors if the contract is broken.
}
\description{
The contract every \code{render_swimlane()} method may rely on.
}
\examples{
validate_swimlane_data(prepare_swimlane(data.frame(s = "a", d = 1), s, d))

}
