% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{validate_waterfall_data}
\alias{validate_waterfall_data}
\title{Validate canonical waterfall data}
\usage{
validate_waterfall_data(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
\code{x}, invisibly. Errors if the contract is broken.
}
\description{
The contract every \code{render_waterfall()} method may rely on, made
machine-checkable. Called at the end of \code{\link[=prepare_waterfall]{prepare_waterfall()}} and again on
entry to \code{\link[=render_waterfall]{render_waterfall()}}.
}
\examples{
validate_waterfall_data(prepare_waterfall(data.frame(i = "a", v = 1), i, v))

}
