% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floaties-package.R
\docType{package}
\name{floaties-package}
\alias{floaties}
\alias{floaties-package}
\title{floaties: Swimlane and Waterfall Plots for Clinical Trial Monitoring}
\description{
Draws the two charts a medical monitor reaches for first while a trial is running: a waterfall of best response, and a swimlane of what happened to each subject over time. Each plot renders with 'ggplot2', 'plotly' or 'echarts4r' from the same call, and takes column roles as arguments so it does not care where the data came from.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vedhav/floaties}
  \item Report bugs at \url{https://github.com/vedhav/floaties/issues}
}

}
\author{
\strong{Maintainer}: Vedha Viyash \email{vedha@appsilon.com}

}
\keyword{internal}
