% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floaties-package.R
\docType{package}
\name{floaties-package}
\alias{floaties}
\alias{floaties-package}
\title{floaties: Swimlane and Waterfall Plots for Multiple Plotting Engines}
\description{
floaties draws swimlane and waterfall plots with the plotting engine of
your choice: ggplot2 for static graphics, or plotly or echarts4r for
interactive htmlwidgets.
}
\details{
Use \code{\link[=swimlane_plot]{swimlane_plot()}} and \code{\link[=waterfall_plot]{waterfall_plot()}} with an \code{engine} argument, or
call a generator such as \code{\link[=swimlane_plotly]{swimlane_plotly()}} directly. Set
\code{options(floaties.engine = "plotly")} to change the default engine for a
session.
}
\author{
\strong{Maintainer}: Vedha Viyash \email{vedha@appsilon.com}

}
\keyword{internal}
