% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floaties-package.R
\docType{package}
\name{floaties-package}
\alias{floaties}
\alias{floaties-package}
\title{floaties: Waterfall and Swimlane Plots for Medical Data Review}
\description{
Create waterfall and swimlane plots for clinical and medical data review using a choice of graphics engines: 'ggplot2', 'plotly', or 'echarts4r'. Functions accept any data frame that follows a simple, documented structure, and columns are supplied with tidy evaluation, so the same code works with 'SDTM', 'ADaM', or any other data standard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vedhav/floaties}
  \item Report bugs at \url{https://github.com/vedhav/floaties/issues}
}

}
\author{
\strong{Maintainer}: Vedha Viyash \email{vedha@appsilon.com} [copyright holder]

}
\keyword{internal}
